# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rooms": {
      "post": {
        "operationId": "Rooms_createRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedRoomResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "create Room"
      },
      "get": {
        "operationId": "Rooms_getRooms",
        "parameters": [
          {
            "name": "usersRoom",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "format": "int32",
              "nullable": true,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Rooms"
      }
    },
    "/rooms/join": {
      "post": {
        "operationId": "Rooms_joinRoom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "join Room"
      }
    },
    "/rooms/{id}": {
      "get": {
        "operationId": "Rooms_getRoomById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Room By Id"
      }
    },
    "/rooms/add-movie/{id}": {
      "post": {
        "operationId": "Rooms_addMovieToRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddRoomMovieDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddedRoomMovieResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "add Movie To Room"
      }
    },
    "/rooms/delete-movie/{id}": {
      "delete": {
        "operationId": "Rooms_removeMovieFromRoom",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveRoomMovieDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "remove Movie From Room"
      }
    },
    "/rooms/{roomId}/rating": {
      "get": {
        "operationId": "Rooms_getRoomRatings",
        "parameters": [
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "nullable": true,
              "type": "string"
            }
          },
          {
            "name": "isWatchTogether",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "rate",
                "userRate"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "sortByUserId",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "format": "int32",
              "nullable": true,
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "format": "int32",
              "nullable": true,
              "default": 10,
              "type": "integer"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 1000000,
              "format": "int32",
              "nullable": true,
              "default": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomRatingsResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Rooms"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Room Ratings"
      }
    },
...
    "/users/me": {
      "get": {
        "operationId": "Users_getCurrentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "get Current User"
      }
    },
    "/users": {
      "post": {
        "operationId": "Users_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [],
        "summary": "create User"
      }
    },
    "/users/{id}": {
      "put": {
        "operationId": "Users_updateUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "update User"
      }
    }
  },
  "info": {
    "title": "Paradiso API",
    "description": "Movie sharing, rooms, ratings and invitations. Successful responses include apiVersion. Use an access token with Authorize; refresh tokens are only for /auth/refresh-tokens.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean",
            "nullable": true
          }
        },
        "required": [
          "name"
        ],
        "additionalProperties": false
      },
      "JoinRoomDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int32",
            "minimum": 1,
            "nullable": true,
            "deprecated": true,
            "description": "Ignored. The authenticated user joins the room."
          },
          "roomId": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        },
        "required": [
          "roomId"
        ],
        "additionalProperties": false
      },
      "AddRoomMovieDto": {
        "type": "object",
        "properties": {
          "dbId": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        },
        "required": [
          "dbId"
        ],
        "additionalProperties": false
      },
      "RemoveRoomMovieDto": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "movieId"
        ],
        "additionalProperties": false
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "writeOnly": true
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "email",
          "password",
          "username"
        ],
        "additionalProperties": false
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string",
            "minLength": 6,
            "writeOnly": true
          },
          "username": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "avatar": {
            "type": "string",
            "maxLength": 255
          }
        },
        "additionalProperties": false
      },
      "PublicUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "username",
          "avatar"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "apiVersion"
        ]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "itemsPerPage": {
            "type": "integer"
          },
          "totalItems": {
            "type": "integer"
          },
          "currentPage": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        },
        "required": [
          "itemsPerPage",
          "totalItems",
          "currentPage",
          "totalPages"
        ]
      },
      "PaginationLinks": {
        "type": "object",
        "properties": {
          "first": {
            "type": "string"
          },
          "previous": {
            "type": "string"
          },
          "next": {
            "type": "string"
          },
          "last": {
            "type": "string"
          },
          "current": {
            "type": "string"
          }
        },
        "required": [
          "first",
          "previous",
          "next",
          "last",
          "current"
        ]
      },
...
      "RoomListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "owner": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "username": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "username"
            ]
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicUser"
            }
          },
          "movies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "type": "string"
                },
                "genres": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "name"
                    ]
                  }
                }
              },
              "required": [
                "id",
                "title",
                "genres"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "image",
          "isPublic",
          "owner",
          "users",
          "movies"
        ]
      },
      "RoomListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomListItem"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "meta",
          "links",
          "apiVersion"
        ]
      },
      "RoomResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "owner": {
            "$ref": "#/components/schemas/PublicUser"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicUser"
            }
          },
          "movies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "type": "string"
                },
                "poster_path": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "title",
                "poster_path"
              ]
            }
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "image",
          "isPublic",
          "owner",
          "users",
          "movies",
          "apiVersion"
        ]
      },
      "CreatedRoomResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "isPublic": {
            "type": "boolean"
          },
          "owner": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "username": {
                "type": "string"
              },
              "avatar": {
                "type": "string",
                "nullable": true
              },
              "role": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "username"
            ]
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "id",
          "name",
          "isPublic",
          "owner",
          "apiVersion"
        ]
      },
      "RoomRating": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "poster_path": {
            "type": "string",
            "nullable": true
          },
          "release_date": {
            "type": "string",
            "nullable": true
          },
          "isWatchedTogether": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "addedBy": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/PublicUser"
              },
              {
                "type": "object",
                "nullable": true,
                "enum": [
                  null
                ]
              }
            ]
          },
          "averageRate": {
            "type": "number",
            "nullable": true
          },
          "userSpecificRate": {
            "type": "number",
            "nullable": true
          },
          "ratings": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "user": {
                  "$ref": "#/components/schemas/PublicUser"
                },
                "rate": {
                  "type": "integer",
                  "nullable": true
                }
              },
              "required": [
                "user",
                "rate"
              ]
            }
          },
          "hasVoted": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "poster_path",
          "release_date",
          "isWatchedTogether",
          "createdAt",
          "addedBy",
          "averageRate",
          "userSpecificRate",
          "ratings",
          "hasVoted"
        ]
      },
      "RoomRatingsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomRating"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "data",
          "meta",
          "links",
          "apiVersion"
        ]
      },
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "apiVersion"
        ]
      },
      "AddedRoomMovieResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "movieId": {
            "type": "string",
            "format": "uuid"
          },
          "apiVersion": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "movieId",
          "apiVersion"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer"
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      }
    }
  }
}